PPSYL-87 - On post select payment, mark order as checkout completed
The form must submit a new hidden input containing the token received on payment onComplete callback from the iframe.
In payment's details, we add a processing status and the token as the payment_id.
Then, payum first call a statusAction, and in that case, we retrieve it instead processing a capture.